Splice supported library to unbreak json schema generation
